--- napi-purple/src/bindings/core.h
@@ -1,8 +1,0 @@
-#ifndef CORE_H_INCLUDED
-#define CORE_H_INCLUDED
-
-#include <node_api.h>
-
-_purple_core_get_version(napi_env env, napi_callback_info info);
-
-#endif
